requirements.txt
numpy==1.24.3
Pillow==10.0.0 